project complete

+++ js/index.js
@@ -4,6 +4,7 @@ let addExerciseSubmit = document.getElementById("add-exercise-submit")
 let showUsername = document.getElementById("showUsername")
 let logOutButton = document.getElementById("logOut-btn")
 let closeExercise = document.getElementById("close-exercise")
+let addExerciseNameError = document.getElementById("add-exercise-name-error");
 let myModal = document.getElementById('exampleModal')
 let myWorkoutLength = 0
 
@@ -11,13 +12,11 @@ let myWorkoutLength = 0
 let getDataCurrentlyLogin = JSON.parse(localStorage.getItem("currentlyLogIn"));
 let getDataAddExercise = localStorage.getItem("exercise") ? JSON.parse(localStorage.getItem("exercise")) : []
 let getUsersData = JSON.parse(localStorage.getItem("users"));
-console.log(getUsersData)
 
 //check if current user is logout
 if (!getDataCurrentlyLogin) {
     location.replace("./template/login.html")
 } else {
-    console.log(getDataCurrentlyLogin.username)
     showUsername.innerHTML += getDataCurrentlyLogin?.username
 }
 
@@ -39,7 +38,6 @@ if (myWorkoutLength) {
 
 // logOut button
 logOutButton.addEventListener("click", () => {
-    console.log("Logout...");
     localStorage.removeItem("currentlyLogIn")
     location.replace("./template/login.html");
 })
@@ -52,21 +50,30 @@ myModal.addEventListener('hidden.bs.modal', function () {
 
 
 //add excercise name 
-console.log(getDataAddExercise);
 addExerciseSubmit.addEventListener("click", () => {
-    let isAlreadyInsert = false;
     let isEmpty = false;
+    let isLetter = true;
+    let isAlreadyInsert = false;
+    
     if (addExercise.value === "") {
         isEmpty = true;
     }
+    //check letter
+    for (let i = 0; i < addExercise.value.length; i++) {
+        let charCode = addExercise.value.charCodeAt(i);
+        if (!(charCode >= 65 && charCode <= 90 || charCode >= 97 && charCode <= 122 || charCode === 32)) {
+            isLetter = false;
+        }
+    }
+    //check alread insert
     for (let i = 0; i < getDataAddExercise.length; i++) {
-        if (getDataAddExercise[i] === addExercise.value.toLowerCase()) {
+        if (getDataAddExercise[i].trim() === addExercise.value.toLowerCase().trim()) {
             isAlreadyInsert = true;
             break
         }
     }
     //add successfully
-    if (!isAlreadyInsert && !isEmpty) {
+    if (!isAlreadyInsert && !isEmpty && isLetter) {
         getDataAddExercise.push(addExercise.value.toLowerCase())
         localStorage.setItem("exercise", JSON.stringify(getDataAddExercise));
         Swal.fire({
@@ -76,20 +83,52 @@ addExerciseSubmit.addEventListener("click", () => {
             showConfirmButton: false,
             timer: 1500
         })
-        console.log(getDataAddExercise);
         addExercise.value = ""
+    } else if (isEmpty) {
+        addExerciseNameError.innerHTML = "Exercise name is required";
+        addExercise.classList.add("border-danger")
+    } else if (isLetter) {
+        addExerciseNameError.innerHTML = "Exercise name is required Letter";
+        addExercise.classList.add("border-danger")
     } else if (isAlreadyInsert) {
         Swal.fire({
             icon: 'error',
             title: 'Oops...',
             text: "Already available",
         })
-    } else if (isEmpty) {
-        Swal.fire({
-            icon: 'error',
-            title: 'Oops...',
-            text: "Empty is not allow",
-        })
+    }
+    else {
+        addExerciseNameError.innerHTML = "";
+        addExercise.classList.remove("border-danger")
     }
 })
 
+//Exercise Name validtion
+addExercise.addEventListener("input",()=> {
+    //check letter
+    let isLetter = true;
+    for (let i = 0; i < addExercise.value.length; i++) {
+        let charCode = addExercise.value.charCodeAt(i);
+        if (!(charCode >= 65 && charCode <= 90 || charCode >= 97 && charCode <= 122)) {
+            isLetter = false;
+        }
+    }
+    if (addExercise.value === "") {
+        addExerciseNameError.innerHTML = "Exercise name is required";
+        addExercise.classList.add("border-danger")
+    }   else if (!isLetter) {
+        addExerciseNameError.innerHTML = "Exercise required only Letters";
+        addExercise.classList.add("border-danger")
+    } 
+    else {
+        addExerciseNameError.innerHTML = ""
+        addExercise.classList.remove("border-danger")
+    }
+})
+
+//close excercise modal button
+myModal.addEventListener('hidden.bs.modal', function () {
+    addExercise.value = ""
+    addExerciseNameError.innerHTML = ""
+    addExercise.classList.remove("border-danger")
+});
